bugfix: doubleのままで表示していた問題を修正、jaotanからのメッセージがゼロでも通知される問題を修正

+++ src/main/java/com/jaoafa/jaoPost/Event/JoinCheck.java
@@ -48,7 +48,7 @@ public class JoinCheck implements Listener {
 				count = res.getInt(1);
 			}
 			if(count != 0){
-				player.sendMessage("[jaoPost] " + ChatColor.GREEN + "受信箱に" + count + "件のメッセージがあります！");
+				player.sendMessage("[jaoPost] " + ChatColor.GREEN + "受信箱に" + ((int)count) + "件のメッセージがあります！");
 			}
 		} catch (SQLException e) {}
 	}
